Refactor: 사용자 인증 및 패키지 변경
- 인증과 관련된 클래스 security 패키지로 이동
- UserPrincipal 클래스 사용
- EmailProvider 를 EmailService로 변경 (패키지도 service로 이동)
