+++ src/main/java/com/zoopick/server/service/AuthService.java
@@ -9,8 +9,7 @@ import com.zoopick.server.exception.BadRequestException;
 import com.zoopick.server.exception.DataNotFoundException;
 import com.zoopick.server.repository.EmailAuthRedisRepository;
 import com.zoopick.server.repository.UserRepository;
-import com.zoopick.server.util.EmailProvider;
-import com.zoopick.server.util.JwtUtil;
+import com.zoopick.server.security.JwtUtil;
 import jakarta.mail.MessagingException;
 import lombok.RequiredArgsConstructor;
 import org.jspecify.annotations.NullMarked;
@@ -32,7 +31,7 @@ public class AuthService {
     private final UserRepository userRepository;
     private final EmailAuthRedisRepository emailAuthRedisRepository;
     private final PasswordEncoder passwordEncoder;
-    private final EmailProvider emailProvider;
+    private final EmailService emailService;
     private final JwtUtil jwtUtil;
     private final TokenValidationService tokenValidationService;
 
@@ -121,7 +120,7 @@ public class AuthService {
         EmailAuth emailAuth = new EmailAuth(email, certificationNumber, createNewExpireTime(), false);
         emailAuthRedisRepository.save(emailAuth);
 
-        emailProvider.senderCertificationMail(email, certificationNumber);
+        emailService.senderCertificationMail(email, certificationNumber);
     }
 
     @Transactional
