+++ src/main/java/com/zoopick/server/service/EmailService.java
@@ -1,4 +1,4 @@
-package com.zoopick.server.util;
+package com.zoopick.server.service;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.util.StreamUtils;
 
 import java.io.IOException;
@@ -17,10 +17,10 @@ import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 
 @Slf4j
-@Component
+@Service
 @RequiredArgsConstructor
 @NullMarked
-public class EmailProvider {
+public class EmailService {
     private static final String SUBJECT = "[명지대 분실물 찾기 서비스] 인증메일 입니다.";
 
     @Value("classpath:/templates/email-certification.html")
